Tests/LibWeb/Text/expected/scrollable-overflow-clipped-child-axis.txt

client width: 100
scroll width: 200
client height: 108
scroll height: 108
